vision-sidecar/pyproject.toml
[project]
name = "strobe-vision"
version = "0.1.0"
requires-python = ">=3.10,<3.13"
dependencies = [
    "torch>=2.0",
    "ultralytics>=8.0",
    "transformers>=4.38,<5.0",
    "Pillow>=10.0",
    "einops>=0.7.0",
    "timm>=0.9.0",
]

[project.scripts]
strobe-vision = "strobe_vision.server:main"
